authenticated user displayed on frontend
this includes removing the alb rules for the admin cognito and making the reviewer cognito into the AuthenticatedSession. Now in the backend we check the group that the user is in before we proceed with the request. We also have a new /user/api namespace which is included in the alb terraform which returns the user's email and groups. Currently this is displayed on unauthenticated pages as Guest. I think I will remove this as it sends a request to /user/api/info each time which causes uneccessary 302 errors and sends a request to the cognito url to try to authorise. there is probably a way around this though. There is support for the backend to return a testing 'dev' user when specifically setting NODE_ENV to development or running make backend from the root dir.